The Concept of Math
It almost does math now. The answer to everything is no longer "42" - it
is now "ADD(four,two,)". On the surface, it does not seem like much of
an improvement, but the important thing is that it interprets things as
mathematical expressions now. It just doesn't do it very well.