Meshin Impossible
I made "mesh" plots, which are a way to visualize vector fields (and
2-dimensional vectors with three inputs, but that takes ridiculous
amounts of time). The code that does so is not pretty, but I'm pretty
proud of it, even if it will only be used for n<3, so a big if statement
would probably work in place of my bajillion for-loops. I also fixed a
couple of bugs.